[Feat] TicketQRCheckViewController 커스텀 메소드 구현 #9
